# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Analytics Platform API
  description: |
    Enterprise web and app analytics API. 
    
    ## Authentication
    
    All endpoints (except public config) require authentication via Bearer token.
    
    - **JWT Token**: Pass in `Authorization: Bearer <access_token>` header
    - **API Token**: Pass in `Authorization: Bearer <api_token>` header (for scoped access)
    
    ## Workspace Context
    
    Most endpoints require a workspace context. Pass the workspace ID in the `X-Workspace-ID` header.
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@analytics.local
servers:
  - url: http://localhost:8080/v1
    description: Local development
  - url: https://api.yourdomain.com/v1
    description: Production
...
paths:
...
  /config/{publicKey}:
    get:
      tags: [Projects]
      summary: Get project configuration (public endpoint)
      description: Returns tracking configuration for the JavaScript SDK. No authentication required.
      parameters:
        - name: publicKey
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectConfig'
        '404':
          description: Project not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ProjectConfig:
      type: object
      properties:
        autotrack:
          type: object
          properties:
            page_views:
              type: boolean
            clicks:
              type: boolean
            forms:
              type: boolean
            scroll_depth:
              type: boolean
            outbound_links:
              type: boolean
            web_vitals:
              type: boolean
            spa_navigation:
              type: boolean
        mask_selectors:
          type: array
          items:
            type: string
        sample_rate:
          type: number
        session_timeout:
          type: integer
        banners:
          type: array
          items:
            type: object
...